refactor: UserModalContents 컴포넌트, model useUserModal 생성
